requirements.txt
moviepy==1.0.3
google-api-python-client==2.157.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
Flask==3.0.2
pyttsx3==2.90
requests==2.31.0
Pillow>=10.0.0
huggingface_hub>=1.9.0
edge-tts>=7.0.0
coqui-tts>=0.24.0
torch>=2.0.0
